Added New Error Checking
Before the binary is ran, it checks and makes sure phantomjs and capture.js
are in the same dir. If the files are not in the dir the program exits with
what is missing. I have also added error checking when running the binary.
If there is an error it gets printed to the screen.